pyproject.toml
[tool.poetry]
name = "mlops_training_pipeline"
authors = [
    "Jae Choi <jae.hyuk.choi.0325@gmail.com>",
]
version = "0.0.1"
packages = [{include = "src"}]
description = "mlops training pipeline"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
kfp = ">=2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = ">=23.9.1"
pylint = "*"
pmdarima = "*"
statsmodels = ">=0.14.3"
scikit-learn = "*"
tensorflow = "==2.17.0"
numpy = "==1.26.4"
pandas = ">=1.4.3"
plotly = "*"
PyYAML = "*"
matplotlib = ">=3.6.0"
google-cloud-bigquery = ">=3.10.0"
google-cloud-aiplatform = "^1.59.0"
pyarrow = "*"

[tool.pytest.ini_options]
markers = [ "mlops_training_pipeline: Mark tests as pipelines tests"]

[tool.coverage.report]
omit = [
  "tests/*",
]